Refactored api_impl.cc and api_impl.h
