+++ src/common/pjrt_implementation/loaded_executable_instance.cc
@@ -1,0 +1,127 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+// This file incorporates work covered by the following copyright and permission
+// notice:
+// SPDX-FileCopyrightText: Copyright 2023 The IREE Authors
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+// https://llvm.org/LICENSE.txt
+
+#include "common/pjrt_implementation/loaded_executable_instance.h"
+
+#include "common/pjrt_implementation/buffer_instance.h"
+#include "common/pjrt_implementation/client_instance.h"
+#include "common/pjrt_implementation/error_instance.h"
+#include "common/pjrt_implementation/utils.h"
+
+namespace tt::pjrt {
+
+LoadedExecutableInstance::~LoadedExecutableInstance() { image_->DecRef(); }
+
+void LoadedExecutableInstance::BindApi(PJRT_Api *api) {
+  api->PJRT_LoadedExecutable_Destroy =
+      +[](PJRT_LoadedExecutable_Destroy_Args *args) -> PJRT_Error * {
+    DLOG_F(LOG_DEBUG,
+           "LoadedExecutableInstance::PJRT_LoadedExecutable_Destroy");
+    delete LoadedExecutableInstance::Unwrap(args->executable);
+    return nullptr;
+  };
+  api->PJRT_LoadedExecutable_AddressableDevices =
+      +[](PJRT_LoadedExecutable_AddressableDevices_Args *args) -> PJRT_Error * {
+    DLOG_F(
+        LOG_DEBUG,
+        "LoadedExecutableInstance::PJRT_LoadedExecutable_AddressableDevices");
+    const std::vector<DeviceInstance *> &devices =
+        LoadedExecutableInstance::Unwrap(args->executable)
+            ->addressable_devices();
+    args->addressable_devices = const_cast<PJRT_Device **>(
+        reinterpret_cast<PJRT_Device *const *>(devices.data()));
+    args->num_addressable_devices = devices.size();
+    return nullptr;
+  };
+  api->PJRT_LoadedExecutable_Delete =
+      +[](PJRT_LoadedExecutable_Delete_Args *args) -> PJRT_Error * {
+    DLOG_F(LOG_DEBUG, "LoadedExecutableInstance::PJRT_LoadedExecutable_Delete");
+    return ErrorInstance::MakeError(tt_pjrt_status::kUnimplemented);
+  };
+  api->PJRT_LoadedExecutable_IsDeleted =
+      +[](PJRT_LoadedExecutable_IsDeleted_Args *args) -> PJRT_Error * {
+    DLOG_F(LOG_DEBUG,
+           "LoadedExecutableInstance::PJRT_LoadedExecutable_IsDeleted_Args");
+    return ErrorInstance::MakeError(tt_pjrt_status::kUnimplemented);
+  };
+  api->PJRT_LoadedExecutable_Execute =
+      +[](PJRT_LoadedExecutable_Execute_Args *args) -> PJRT_Error * {
+    DLOG_F(LOG_DEBUG,
+           "LoadedExecutableInstance::PJRT_LoadedExecutable_Execute");
+    return ErrorInstance::MakeError(
+        LoadedExecutableInstance::Unwrap(args->executable)->Execute(args));
+  };
+  api->PJRT_LoadedExecutable_GetExecutable =
+      +[](PJRT_LoadedExecutable_GetExecutable_Args *args) -> PJRT_Error * {
+    DLOG_F(LOG_DEBUG,
+           "LoadedExecutableInstance::PJRT_LoadedExecutable_GetExecutable");
+    LoadedExecutableInstance *loaded_exe =
+        LoadedExecutableInstance::Unwrap(args->loaded_executable);
+    ExecutableImage *image = loaded_exe->image_;
+
+    image->AddRef();
+    args->executable = *image;
+    return nullptr;
+  };
+}
+
+tt_pjrt_status
+LoadedExecutableInstance::Execute(PJRT_LoadedExecutable_Execute_Args *args) {
+  DLOG_F(LOG_DEBUG, "LoadedExecutableInstance::Execute");
+
+  auto [system_desc, chip_ids] = tt::runtime::getCurrentSystemDesc();
+  int dev_0 = chip_ids[0];
+  tt::runtime::Device device = tt::runtime::openDevice({dev_0});
+
+  assert(args->num_devices == 1);
+  int dev_index = 0;
+  tt::runtime::Binary binary(image_->get_binary());
+
+  std::vector<tt::runtime::Tensor> rt_inputs;
+  rt_inputs.reserve(args->num_args);
+
+  for (size_t i = 0; i < args->num_args; ++i) {
+    BufferInstance *buffer =
+        BufferInstance::Unwrap(args->argument_lists[dev_index][i]);
+    rt_inputs.emplace_back(buffer->tensor());
+    DLOG_F(INFO, "Runtime input id: %d", buffer->unique_id());
+  }
+
+  std::vector<tt::runtime::Tensor> rt_outputs =
+      tt::runtime::submit(device, binary, 0, rt_inputs);
+  std::vector<tt::runtime::TensorDesc> output_specs =
+      binary.getProgramOutputs(0);
+
+  assert(rt_outputs.size() == output_specs.size());
+
+  for (size_t i = 0; i < output_specs.size(); ++i) {
+    bool is_scalar = client_.isOutputScalar(i);
+    // PJRT expects an empty shape for scalars.
+    std::vector<std::uint32_t> output_shape =
+        is_scalar ? std::vector<std::uint32_t>() : output_specs[i].shape;
+    auto result_buffer = std::make_unique<BufferInstance>(
+        *this->addressable_devices_[dev_index], rt_outputs[i], output_shape,
+        output_specs[i].stride);
+    result_buffer->setType(tt::pjrt::utils::convertElementTypeToBufferType(
+        output_specs[i].dataType));
+    DLOG_F(INFO, "Runtime output id: %d", result_buffer->unique_id());
+    args->output_lists[dev_index][i] = *(result_buffer.release());
+  }
+
+  if (args->device_complete_events) {
+    args->device_complete_events[dev_index] = *(new EventInstance());
+  }
+
+  tt::runtime::closeDevice(device);
+
+  return tt_pjrt_status::kSuccess;
+}
+
+} // namespace tt::pjrt
